refactor: rename function and move it into separate folder
